Isolate the query module
